// src/interfaces/models/IGasoline.ts
export interface Igasoline {
  id: number;
  motorcicle_id: number;
  board: string;
  date: string;
  date_formart: string;
  km_last: number;
  km: number;
  km_per_run: number;
  km_per_liters: number;
  liters: number;
  price: number;
  total: number;
  created_at: string;
  updated_at: string;
}